Add workflow task for syncing greenhouse postings
Added a new "Sync Greenhouse Postings" workflow task that handles
fetching greenhouse job postings via the `job_posts` endpoint and
create/update content items within Orchard Core.

To assist with the sync process, a new index table has been created
which stores the ID of the job post within Greenhouse and when the job
posting was updated. Each sync will fetch the latest updated date from
the index and use it to only fetch job posts that have been updated
after the date. When handling the response, a query is made to see if
any content items are already associated to the Greenhouse job posting
ID, if so then the content item is updated, otherwise a new content item
is created.

https://developers.greenhouse.io/harvest.html#get-list-job-posts