Add configuration options to sync task
Added a couple of new options that can be configured when editing the
sync task in a workflow. Firstly, "Author" can be used to define who is
set as the author when a content item is created/updated. "Content Type"
gives the ability to have a different content type from the default one
that is added ("Greenhouse Posting").