Handle application form submission
- Add model binding which handles taking values submitted in the request
form values and converting it to an object that can be used as the
payload in the request to the Greenhouse API.
- Add additional API requests that submits attachments that should be
associated with the candidates application.
- Add form validation, ensuring that there is both client & server side
validation which cater for the different types of fields.
- When either there is a validation issue caught on the server side or
an issue happens during binding/valdation/API submission, ensure the
user is redirected to the form with their submitted values intact. This
is done by transfering the `ModelState` to `TempData`, which can be
accessed when rendering the form fields. In order to gain access to
`TempData`, the templates used for the different field types have been
converted from Liquid to Razor.