Add new part for managing content on job application form
Introduced a new `GreenhousePostingFormPart` which will be used to
contain configuration properties for the form. This change only
introduces the ability to toggle whether the application form is
displayed or not (default to display). This new part will also take
responsibility for display the form when rendering a content item.