Add option to sync to prevent multiple postings for a single job
Job's in Greenhouse can have multiple job postings associated to it. It
could be desirable that instead of showing multiple postings for the
same job, visitors to the site only see a single posting for a job to
avoid seeing the same posting over and over. When setting up the sync
task, users can now select to ignore postings associated to a job that
is already represented by a content item.