Update sync job board activity to use job board API to fetch job posts
Instead of using the harvest API to fetch jobs returned on the job
board, the job board API contains an endpoint for fetching a single job.
This doesn't contain all the same information as a job posting via the
harvest API, however it contains everything needed to display a job on
the front-end (title, location, department, etc...).