Refactor apply endpoint to pass contentItemId in URL instead of referrer
Controller action was using the URL referrer to obtain the content item
that represents the job that is being applied for. However, in a proxy
environment the referrer might not be the page the user has submitted
from. To get round it, the content item ID is now passed in the apply
URL (so it's `/greenhouse/apply/1234`).