Improve handling of different metadata fields
Within Greenhouse it's possible to have metadata fields that have a
value of a single string or a collection of strings. This means the data
in the API response from Greenhouse can be either a string or an array
of strings. This change adds a converter to handle this scenario and
converts any single string values to a collection of strings.

Additionally, this change updates the GreenhousePostingPartIndexHandler
to clean up the name of the meta data field that's added to the index
document.